Add Prem's security reporting talk proposal
We all have been through this; Someone uploaded a exploited gem to
RubyGems just to show that RubyGems has a security issue, someone posted
an issue on GitHub just to get an attention from the core team members,
someone post a link to a Gist on how to exploited something on Hacker
News just to prove that it's possible, or someone just cut a security
release without a proper security advisory on Friday evening when
everybody left their work.

As a community, I believe that we can be better than this. We can all
learn together on how to do security reporting and disclosing in a right
way. Be responsible, polite, and respectful to everyone so that we're
not ruining anyone's day when a vulnerability has been exposed.