Add Prem's open source contributing proposal
It has been a while since I started contributing to open source, and I
have to say that the starting process is not really that smooth. Every
time I'm giving a talk at several places, people was keep asking me on
how I started contributing to an open source. I believe There are a lot
of things you need to do to get started, and you might probably get
tired of it and give up if you didn't get it right the first time. I'd
like this talk to encourage and inspire more people to contribute to an
open source project, by giving them the idea on how to get started and
how it should be done.

Also, as a maintainer of two gems, I think I would like to spread the
word as well on how to lower the barrier to contribution. That way, new
people could easily get started when they wanted to contribute back to
the community, like I did.